core: Define temporary path for filesystem artifacts
This is required because in some cases there are intermediate files between
the source files and the target files. The target files need to be added to
the filesystem image, but the source and intermediate files don't. With a
temporary path, intermediate files can be saved there.

This is needed for GL2D sprite sets stored in NitroFS, which will be
implemented in a future commit.