core: Fix linker invocation for C++
This commit is equivalent to the following commit of the BlocksDS SDK
repository:

https://github.com/blocksds/sdk/commit/afb89fede0fff150656c8ccbc9e4603b9b85dfe2

Using g++ to link object files automatically links the STD of C++. It looks
like this library is linked outside of the start/end-group commands with
the other libraries, which causes undefined reference errors when linking
C++ projects that use standard library functions that interface with the
OS, like `fopen()`.

Thanks to @lifehackerhansol for the report and testing the fix.